hyde v0.0.1 Hyde

Basic (Redis Backed) Feature Toggling

Activate/Deactivate based on feature name & optionally an identifying integer to compare against.

Requires an initialized ExRedis client to be passed into the function calls

Functions

activate(client, feature, user_id \\ 0)

Parameters

client client to connect to redis

feature Atom that addresses the feature

user_id Optional: Integer that represents the identity of a user

Examples

iex> {:ok, client} = Exredis.start_link
iex> client |> Hyde.activate(:my_feature)
true

active?(client, feature, user_id \\ 0)

Parameters

client client to connect to redis

feature Atom that represents the name of the feature

user_id Optional: Integer that represents the identity of a user

Examples

iex> {:ok, client} = Exredis.start_link; 
iex> client |> Hyde.activate(:my_active_feature)
iex> client |> Hyde.active?(:my_active_feature)
true

deactivate(client, feature, user_id \\ 0)

Parameters

client client to connect to redis

feature Atom that represents the name of the feature

user_id Optional: Integer that represents the identity of a user

Examples

iex> {:ok, client} = Exredis.start_link
iex> client |> Hyde.deactivate(:my_feature)
true

inactive?(client, feature, user_id \\ 0)

Parameters

client client to connect to redis

feature Atom that represents the name of the feature

user_id Optional: Integer that represents the identity of a user

Examples

iex> {:ok, client} = Exredis.start_link; 
iex> client |> Hyde.inactive?(:my_inactive_feature)
true